lib, pimd, sharpd: Various output string cleanups
Various compilers in our CI system were complaining about various
auto-conversions.  Let's get these cleaned up a bit more.

+++ sharpd/sharp_vty.c
@@ -150,7 +150,7 @@ DEFPY (install_routes_data_dump,
 		sg.r.total_routes,
 		sg.r.installed_routes,
 		sg.r.removed_routes,
-		r.tv_sec, r.tv_usec);
+		r.tv_sec, (long int)r.tv_usec);
 
 	return CMD_SUCCESS;
 }
